refactor: move batcher.py and config.py to the plugin_runner module
